make requested changes and move time limited max size cache to gear
