+++ gear/gear/time_limited_max_size_cache.py
